Make rom match checksum

+++ src/field_map_obj_helpers.c
@@ -76,26 +76,26 @@ void Step1(struct Sprite *sprite, u8 dir)
 
 void Step2(struct Sprite *sprite, u8 dir)
 {
-    sprite->pos1.x += 2 * gDirectionToVector[dir].x;
-    sprite->pos1.y += 2 * gDirectionToVector[dir].y;
+    sprite->pos1.x += 2 * (u16) gDirectionToVector[dir].x;
+    sprite->pos1.y += 2 * (u16) gDirectionToVector[dir].y;
 }
 
 void Step3(struct Sprite *sprite, u8 dir)
 {
-    sprite->pos1.x += 2 * gDirectionToVector[dir].x + gDirectionToVector[dir].x;
-    sprite->pos1.y += 2 * gDirectionToVector[dir].y + gDirectionToVector[dir].y;
+    sprite->pos1.x += 2 * (u16) gDirectionToVector[dir].x + (u16) gDirectionToVector[dir].x;
+    sprite->pos1.y += 2 * (u16) gDirectionToVector[dir].y + (u16) gDirectionToVector[dir].y;
 }
 
 void Step4(struct Sprite *sprite, u8 dir)
 {
-    sprite->pos1.x += 4 * gDirectionToVector[dir].x;
-    sprite->pos1.y += 4 * gDirectionToVector[dir].y;
+    sprite->pos1.x += 4 * (u16) gDirectionToVector[dir].x;
+    sprite->pos1.y += 4 * (u16) gDirectionToVector[dir].y;
 }
 
 void Step8(struct Sprite *sprite, u8 dir)
 {
-    sprite->pos1.x += 8 * gDirectionToVector[dir].x;
-    sprite->pos1.y += 8 * gDirectionToVector[dir].y;
+    sprite->pos1.x += 8 * (u16) gDirectionToVector[dir].x;
+    sprite->pos1.y += 8 * (u16) gDirectionToVector[dir].y;
 }
 
 void oamt_npc_ministep_reset(struct Sprite *sprite, u8 a2, u8 a3)
